sstable: split block.go
Split block.go into `block_writer.go`, `block_iter.go`, and
`block_fragment_iter.go`. This will allow easier code generation for
the streamlined version of `blockIter`.
