+++ sstable/block_iter.go
@@ -8,289 +8,14 @@ import (
 	"bytes"
 	"context"
 	"encoding/binary"
-	"slices"
 	"unsafe"
 
 	"github.com/cockroachdb/errors"
 	"github.com/cockroachdb/pebble/internal/base"
 	"github.com/cockroachdb/pebble/internal/invariants"
-	"github.com/cockroachdb/pebble/internal/keyspan"
 	"github.com/cockroachdb/pebble/internal/manual"
-	"github.com/cockroachdb/pebble/internal/rangedel"
-	"github.com/cockroachdb/pebble/internal/rangekey"
+	"golang.org/x/exp/slices"
 )
-
-func uvarintLen(v uint32) int {
-	i := 0
-	for v >= 0x80 {
-		v >>= 7
-		i++
-	}
-	return i + 1
-}
-
-type blockWriter struct {
-	restartInterval int
-	nEntries        int
-	nextRestart     int
-	buf             []byte
-	// For datablocks in TableFormatPebblev3, we steal the most significant bit
-	// in restarts for encoding setHasSameKeyPrefixSinceLastRestart. This leaves
-	// us with 31 bits, which is more than enough (no one needs > 2GB blocks).
-	// Typically, restarts occur every 16 keys, and by storing this bit with the
-	// restart, we can optimize for the case where a user wants to skip to the
-	// next prefix which happens to be in the same data block, but is > 16 keys
-	// away. We have seen production situations with 100+ versions per MVCC key
-	// (which share the same prefix). Additionally, for such writers, the prefix
-	// compression of the key, that shares the key with the preceding key, is
-	// limited to the prefix part of the preceding key -- this ensures that when
-	// doing NPrefix (see blockIter) we don't need to assemble the full key
-	// for each step since by limiting the length of the shared key we are
-	// ensuring that any of the keys with the same prefix can be used to
-	// assemble the full key when the prefix does change.
-	restarts []uint32
-	// Do not read curKey directly from outside blockWriter since it can have
-	// the InternalKeyKindSSTableInternalObsoleteBit set. Use getCurKey() or
-	// getCurUserKey() instead.
-	curKey []byte
-	// curValue excludes the optional prefix provided to
-	// storeWithOptionalValuePrefix.
-	curValue []byte
-	prevKey  []byte
-	tmp      [4]byte
-	// We don't know the state of the sets that were at the end of the previous
-	// block, so this is initially 0. It may be true for the second and later
-	// restarts in a block. Not having inter-block information is fine since we
-	// will optimize by stepping through restarts only within the same block.
-	// Note that the first restart is the first key in the block.
-	setHasSameKeyPrefixSinceLastRestart bool
-}
-
-func (w *blockWriter) clear() {
-	*w = blockWriter{
-		buf:      w.buf[:0],
-		restarts: w.restarts[:0],
-		curKey:   w.curKey[:0],
-		curValue: w.curValue[:0],
-		prevKey:  w.prevKey[:0],
-	}
-}
-
-// MaximumBlockSize is an extremely generous maximum block size of 256MiB. We
-// explicitly place this limit to reserve a few bits in the restart for
-// internal use.
-const MaximumBlockSize = 1 << 28
-const setHasSameKeyPrefixRestartMask uint32 = 1 << 31
-const restartMaskLittleEndianHighByteWithoutSetHasSamePrefix byte = 0b0111_1111
-const restartMaskLittleEndianHighByteOnlySetHasSamePrefix byte = 0b1000_0000
-
-func (w *blockWriter) getCurKey() InternalKey {
-	k := base.DecodeInternalKey(w.curKey)
-	k.Trailer = k.Trailer & trailerObsoleteMask
-	return k
-}
-
-func (w *blockWriter) getCurUserKey() []byte {
-	n := len(w.curKey) - base.InternalTrailerLen
-	if n < 0 {
-		panic(errors.AssertionFailedf("corrupt key in blockWriter buffer"))
-	}
-	return w.curKey[:n:n]
-}
-
-// If !addValuePrefix, the valuePrefix is ignored.
-func (w *blockWriter) storeWithOptionalValuePrefix(
-	keySize int,
-	value []byte,
-	maxSharedKeyLen int,
-	addValuePrefix bool,
-	valuePrefix valuePrefix,
-	setHasSameKeyPrefix bool,
-) {
-	shared := 0
-	if !setHasSameKeyPrefix {
-		w.setHasSameKeyPrefixSinceLastRestart = false
-	}
-	if w.nEntries == w.nextRestart {
-		w.nextRestart = w.nEntries + w.restartInterval
-		restart := uint32(len(w.buf))
-		if w.setHasSameKeyPrefixSinceLastRestart {
-			restart = restart | setHasSameKeyPrefixRestartMask
-		}
-		w.setHasSameKeyPrefixSinceLastRestart = true
-		w.restarts = append(w.restarts, restart)
-	} else {
-		// TODO(peter): Manually inlined version of base.SharedPrefixLen(). This
-		// is 3% faster on BenchmarkWriter on go1.16. Remove if future versions
-		// show this to not be a performance win. For now, functions that use of
-		// unsafe cannot be inlined.
-		n := maxSharedKeyLen
-		if n > len(w.prevKey) {
-			n = len(w.prevKey)
-		}
-		asUint64 := func(b []byte, i int) uint64 {
-			return binary.LittleEndian.Uint64(b[i:])
-		}
-		for shared < n-7 && asUint64(w.curKey, shared) == asUint64(w.prevKey, shared) {
-			shared += 8
-		}
-		for shared < n && w.curKey[shared] == w.prevKey[shared] {
-			shared++
-		}
-	}
-
-	lenValuePlusOptionalPrefix := len(value)
-	if addValuePrefix {
-		lenValuePlusOptionalPrefix++
-	}
-	needed := 3*binary.MaxVarintLen32 + len(w.curKey[shared:]) + lenValuePlusOptionalPrefix
-	n := len(w.buf)
-	if cap(w.buf) < n+needed {
-		newCap := 2 * cap(w.buf)
-		if newCap == 0 {
-			newCap = 1024
-		}
-		for newCap < n+needed {
-			newCap *= 2
-		}
-		newBuf := make([]byte, n, newCap)
-		copy(newBuf, w.buf)
-		w.buf = newBuf
-	}
-	w.buf = w.buf[:n+needed]
-
-	// TODO(peter): Manually inlined versions of binary.PutUvarint(). This is 15%
-	// faster on BenchmarkWriter on go1.13. Remove if go1.14 or future versions
-	// show this to not be a performance win.
-	{
-		x := uint32(shared)
-		for x >= 0x80 {
-			w.buf[n] = byte(x) | 0x80
-			x >>= 7
-			n++
-		}
-		w.buf[n] = byte(x)
-		n++
-	}
-
-	{
-		x := uint32(keySize - shared)
-		for x >= 0x80 {
-			w.buf[n] = byte(x) | 0x80
-			x >>= 7
-			n++
-		}
-		w.buf[n] = byte(x)
-		n++
-	}
-
-	{
-		x := uint32(lenValuePlusOptionalPrefix)
-		for x >= 0x80 {
-			w.buf[n] = byte(x) | 0x80
-			x >>= 7
-			n++
-		}
-		w.buf[n] = byte(x)
-		n++
-	}
-
-	n += copy(w.buf[n:], w.curKey[shared:])
-	if addValuePrefix {
-		w.buf[n : n+1][0] = byte(valuePrefix)
-		n++
-	}
-	n += copy(w.buf[n:], value)
-	w.buf = w.buf[:n]
-
-	w.curValue = w.buf[n-len(value):]
-
-	w.nEntries++
-}
-
-func (w *blockWriter) add(key InternalKey, value []byte) {
-	w.addWithOptionalValuePrefix(
-		key, false, value, len(key.UserKey), false, 0, false)
-}
-
-// Callers that always set addValuePrefix to false should use add() instead.
-//
-// isObsolete indicates whether this key-value pair is obsolete in this
-// sstable (only applicable when writing data blocks) -- see the comment in
-// table.go and the longer one in format.go. addValuePrefix adds a 1 byte
-// prefix to the value, specified in valuePrefix -- this is used for data
-// blocks in TableFormatPebblev3 onwards for SETs (see the comment in
-// format.go, with more details in value_block.go). setHasSameKeyPrefix is
-// also used in TableFormatPebblev3 onwards for SETs.
-func (w *blockWriter) addWithOptionalValuePrefix(
-	key InternalKey,
-	isObsolete bool,
-	value []byte,
-	maxSharedKeyLen int,
-	addValuePrefix bool,
-	valuePrefix valuePrefix,
-	setHasSameKeyPrefix bool,
-) {
-	w.curKey, w.prevKey = w.prevKey, w.curKey
-
-	size := key.Size()
-	if cap(w.curKey) < size {
-		w.curKey = make([]byte, 0, size*2)
-	}
-	w.curKey = w.curKey[:size]
-	if isObsolete {
-		key.Trailer = key.Trailer | trailerObsoleteBit
-	}
-	key.Encode(w.curKey)
-
-	w.storeWithOptionalValuePrefix(
-		size, value, maxSharedKeyLen, addValuePrefix, valuePrefix, setHasSameKeyPrefix)
-}
-
-func (w *blockWriter) finish() []byte {
-	// Write the restart points to the buffer.
-	if w.nEntries == 0 {
-		// Every block must have at least one restart point.
-		if cap(w.restarts) > 0 {
-			w.restarts = w.restarts[:1]
-			w.restarts[0] = 0
-		} else {
-			w.restarts = append(w.restarts, 0)
-		}
-	}
-	tmp4 := w.tmp[:4]
-	for _, x := range w.restarts {
-		binary.LittleEndian.PutUint32(tmp4, x)
-		w.buf = append(w.buf, tmp4...)
-	}
-	binary.LittleEndian.PutUint32(tmp4, uint32(len(w.restarts)))
-	w.buf = append(w.buf, tmp4...)
-	result := w.buf
-
-	// Reset the block state.
-	w.nEntries = 0
-	w.nextRestart = 0
-	w.buf = w.buf[:0]
-	w.restarts = w.restarts[:0]
-	return result
-}
-
-// emptyBlockSize holds the size of an empty block. Every block ends
-// in a uint32 trailer encoding the number of restart points within the
-// block.
-const emptyBlockSize = 4
-
-func (w *blockWriter) estimatedSize() int {
-	return len(w.buf) + 4*len(w.restarts) + emptyBlockSize
-}
-
-type blockEntry struct {
-	offset   int32
-	keyStart int32
-	keyEnd   int32
-	valStart int32
-	valSize  int32
-}
 
 // blockIter is an iterator over a single block of data.
 //
@@ -460,6 +185,14 @@ type blockIter struct {
 	}
 	synthSuffixBuf            []byte
 	firstUserKeyWithPrefixBuf []byte
+}
+
+type blockEntry struct {
+	offset   int32
+	keyStart int32
+	keyEnd   int32
+	valStart int32
+	valSize  int32
 }
 
 // blockIter implements the base.InternalIterator interface.
@@ -1804,318 +1537,3 @@ func (i *blockIter) SetContext(_ context.Context) {}
 func (i *blockIter) valid() bool {
 	return i.offset >= 0 && i.offset < i.restarts
 }
-
-// fragmentBlockIter wraps a blockIter, implementing the
-// keyspan.FragmentIterator interface. It's used for reading range deletion and
-// range key blocks.
-//
-// Range deletions and range keys are fragmented before they're persisted to the
-// block. Overlapping fragments have identical bounds.  The fragmentBlockIter
-// gathers all the fragments with identical bounds within a block and returns a
-// single keyspan.Span describing all the keys defined over the span.
-//
-// # Memory lifetime
-//
-// A Span returned by fragmentBlockIter is only guaranteed to be stable until
-// the next fragmentBlockIter iteration positioning method. A Span's Keys slice
-// may be reused, so the user must not assume it's stable.
-//
-// Blocks holding range deletions and range keys are configured to use a restart
-// interval of 1. This provides key stability. The caller may treat the various
-// byte slices (start, end, suffix, value) as stable for the lifetime of the
-// iterator.
-type fragmentBlockIter struct {
-	blockIter blockIter
-	keyBuf    [2]keyspan.Key
-	span      keyspan.Span
-	dir       int8
-	closeHook func(i keyspan.FragmentIterator) error
-
-	// elideSameSeqnum, if true, returns only the first-occurring (in forward
-	// order) Key for each sequence number.
-	elideSameSeqnum bool
-}
-
-func (i *fragmentBlockIter) resetForReuse() fragmentBlockIter {
-	return fragmentBlockIter{blockIter: i.blockIter.resetForReuse()}
-}
-
-func (i *fragmentBlockIter) decodeSpanKeys(k *InternalKey, internalValue []byte) error {
-	// TODO(jackson): The use of i.span.Keys to accumulate keys across multiple
-	// calls to Decode is too confusing and subtle. Refactor to make it
-	// explicit.
-
-	// decode the contents of the fragment's value. This always includes at
-	// least the end key: RANGEDELs store the end key directly as the value,
-	// whereas the various range key kinds store are more complicated.  The
-	// details of the range key internal value format are documented within the
-	// internal/rangekey package.
-	var err error
-	switch k.Kind() {
-	case base.InternalKeyKindRangeDelete:
-		i.span = rangedel.Decode(*k, internalValue, i.span.Keys)
-	case base.InternalKeyKindRangeKeySet, base.InternalKeyKindRangeKeyUnset, base.InternalKeyKindRangeKeyDelete:
-		i.span, err = rangekey.Decode(*k, internalValue, i.span.Keys)
-	default:
-		i.span = keyspan.Span{}
-		err = base.CorruptionErrorf("pebble: corrupt keyspan fragment of kind %d", k.Kind())
-	}
-	return err
-}
-
-func (i *fragmentBlockIter) elideKeysOfSameSeqNum() {
-	if invariants.Enabled {
-		if !i.elideSameSeqnum || len(i.span.Keys) == 0 {
-			panic("elideKeysOfSameSeqNum called when it should not be")
-		}
-	}
-	lastSeqNum := i.span.Keys[0].SeqNum()
-	k := 1
-	for j := 1; j < len(i.span.Keys); j++ {
-		if lastSeqNum != i.span.Keys[j].SeqNum() {
-			lastSeqNum = i.span.Keys[j].SeqNum()
-			i.span.Keys[k] = i.span.Keys[j]
-			k++
-		}
-	}
-	i.span.Keys = i.span.Keys[:k]
-}
-
-// gatherForward gathers internal keys with identical bounds. Keys defined over
-// spans of the keyspace are fragmented such that any overlapping key spans have
-// identical bounds. When these spans are persisted to a range deletion or range
-// key block, they may be persisted as multiple internal keys in order to encode
-// multiple sequence numbers or key kinds.
-//
-// gatherForward iterates forward, re-combining the fragmented internal keys to
-// reconstruct a keyspan.Span that holds all the keys defined over the span.
-func (i *fragmentBlockIter) gatherForward(
-	k *InternalKey, lazyValue base.LazyValue,
-) (*keyspan.Span, error) {
-	i.span = keyspan.Span{}
-	if k == nil || !i.blockIter.valid() {
-		return nil, nil
-	}
-	// Use the i.keyBuf array to back the Keys slice to prevent an allocation
-	// when a span contains few keys.
-	i.span.Keys = i.keyBuf[:0]
-
-	// Decode the span's end key and individual keys from the value.
-	internalValue := lazyValue.InPlaceValue()
-	if err := i.decodeSpanKeys(k, internalValue); err != nil {
-		return nil, err
-	}
-	prevEnd := i.span.End
-
-	// There might exist additional internal keys with identical bounds encoded
-	// within the block. Iterate forward, accumulating all the keys with
-	// identical bounds to s.
-	k, lazyValue = i.blockIter.Next()
-	internalValue = lazyValue.InPlaceValue()
-	for k != nil && i.blockIter.cmp(k.UserKey, i.span.Start) == 0 {
-		if err := i.decodeSpanKeys(k, internalValue); err != nil {
-			return nil, err
-		}
-
-		// Since k indicates an equal start key, the encoded end key must
-		// exactly equal the original end key from the first internal key.
-		// Overlapping fragments are required to have exactly equal start and
-		// end bounds.
-		if i.blockIter.cmp(prevEnd, i.span.End) != 0 {
-			i.span = keyspan.Span{}
-			return nil, base.CorruptionErrorf("pebble: corrupt keyspan fragmentation")
-		}
-		k, lazyValue = i.blockIter.Next()
-		internalValue = lazyValue.InPlaceValue()
-	}
-	if i.elideSameSeqnum && len(i.span.Keys) > 0 {
-		i.elideKeysOfSameSeqNum()
-	}
-	// i.blockIter is positioned over the first internal key for the next span.
-	return &i.span, nil
-}
-
-// gatherBackward gathers internal keys with identical bounds. Keys defined over
-// spans of the keyspace are fragmented such that any overlapping key spans have
-// identical bounds. When these spans are persisted to a range deletion or range
-// key block, they may be persisted as multiple internal keys in order to encode
-// multiple sequence numbers or key kinds.
-//
-// gatherBackward iterates backwards, re-combining the fragmented internal keys
-// to reconstruct a keyspan.Span that holds all the keys defined over the span.
-func (i *fragmentBlockIter) gatherBackward(
-	k *InternalKey, lazyValue base.LazyValue,
-) (*keyspan.Span, error) {
-	i.span = keyspan.Span{}
-	if k == nil || !i.blockIter.valid() {
-		return nil, nil
-	}
-	// Use the i.keyBuf array to back the Keys slice to prevent an allocation
-	// when a span contains few keys.
-	i.span.Keys = i.keyBuf[:0]
-
-	// Decode the span's end key and individual keys from the value.
-	internalValue := lazyValue.InPlaceValue()
-	if err := i.decodeSpanKeys(k, internalValue); err != nil {
-		return nil, err
-	}
-	prevEnd := i.span.End
-
-	// There might exist additional internal keys with identical bounds encoded
-	// within the block. Iterate backward, accumulating all the keys with
-	// identical bounds to s.
-	k, lazyValue = i.blockIter.Prev()
-	internalValue = lazyValue.InPlaceValue()
-	for k != nil && i.blockIter.cmp(k.UserKey, i.span.Start) == 0 {
-		if err := i.decodeSpanKeys(k, internalValue); err != nil {
-			return nil, err
-		}
-
-		// Since k indicates an equal start key, the encoded end key must
-		// exactly equal the original end key from the first internal key.
-		// Overlapping fragments are required to have exactly equal start and
-		// end bounds.
-		if i.blockIter.cmp(prevEnd, i.span.End) != 0 {
-			i.span = keyspan.Span{}
-			return nil, base.CorruptionErrorf("pebble: corrupt keyspan fragmentation")
-		}
-		k, lazyValue = i.blockIter.Prev()
-		internalValue = lazyValue.InPlaceValue()
-	}
-	// i.blockIter is positioned over the last internal key for the previous
-	// span.
-
-	// Backwards iteration encounters internal keys in the wrong order.
-	keyspan.SortKeysByTrailer(&i.span.Keys)
-
-	if i.elideSameSeqnum && len(i.span.Keys) > 0 {
-		i.elideKeysOfSameSeqNum()
-	}
-	return &i.span, nil
-}
-
-// Close implements (keyspan.FragmentIterator).Close.
-func (i *fragmentBlockIter) Close() error {
-	var err error
-	if i.closeHook != nil {
-		err = i.closeHook(i)
-	}
-	err = firstError(err, i.blockIter.Close())
-	return err
-}
-
-// First implements (keyspan.FragmentIterator).First
-func (i *fragmentBlockIter) First() (*keyspan.Span, error) {
-	i.dir = +1
-	return i.gatherForward(i.blockIter.First())
-}
-
-// Last implements (keyspan.FragmentIterator).Last.
-func (i *fragmentBlockIter) Last() (*keyspan.Span, error) {
-	i.dir = -1
-	return i.gatherBackward(i.blockIter.Last())
-}
-
-// Next implements (keyspan.FragmentIterator).Next.
-func (i *fragmentBlockIter) Next() (*keyspan.Span, error) {
-	switch {
-	case i.dir == -1 && !i.span.Valid():
-		// Switching directions.
-		//
-		// i.blockIter is exhausted, before the first key. Move onto the first.
-		i.blockIter.First()
-		i.dir = +1
-	case i.dir == -1 && i.span.Valid():
-		// Switching directions.
-		//
-		// i.blockIter is currently positioned over the last internal key for
-		// the previous span. Next it once to move to the first internal key
-		// that makes up the current span, and gatherForwaad to land on the
-		// first internal key making up the next span.
-		//
-		// In the diagram below, if the last span returned to the user during
-		// reverse iteration was [b,c), i.blockIter is currently positioned at
-		// [a,b). The block iter must be positioned over [d,e) to gather the
-		// next span's fragments.
-		//
-		//    ... [a,b) [b,c) [b,c) [b,c) [d,e) ...
-		//          ^                       ^
-		//     i.blockIter                 want
-		if x, err := i.gatherForward(i.blockIter.Next()); err != nil {
-			return nil, err
-		} else if invariants.Enabled && !x.Valid() {
-			panic("pebble: invariant violation: next entry unexpectedly invalid")
-		}
-		i.dir = +1
-	}
-	// We know that this blockIter has in-place values.
-	return i.gatherForward(&i.blockIter.ikey, base.MakeInPlaceValue(i.blockIter.val))
-}
-
-// Prev implements (keyspan.FragmentIterator).Prev.
-func (i *fragmentBlockIter) Prev() (*keyspan.Span, error) {
-	switch {
-	case i.dir == +1 && !i.span.Valid():
-		// Switching directions.
-		//
-		// i.blockIter is exhausted, after the last key. Move onto the last.
-		i.blockIter.Last()
-		i.dir = -1
-	case i.dir == +1 && i.span.Valid():
-		// Switching directions.
-		//
-		// i.blockIter is currently positioned over the first internal key for
-		// the next span. Prev it once to move to the last internal key that
-		// makes up the current span, and gatherBackward to land on the last
-		// internal key making up the previous span.
-		//
-		// In the diagram below, if the last span returned to the user during
-		// forward iteration was [b,c), i.blockIter is currently positioned at
-		// [d,e). The block iter must be positioned over [a,b) to gather the
-		// previous span's fragments.
-		//
-		//    ... [a,b) [b,c) [b,c) [b,c) [d,e) ...
-		//          ^                       ^
-		//        want                  i.blockIter
-		if x, err := i.gatherBackward(i.blockIter.Prev()); err != nil {
-			return nil, err
-		} else if invariants.Enabled && !x.Valid() {
-			panic("pebble: invariant violation: previous entry unexpectedly invalid")
-		}
-		i.dir = -1
-	}
-	// We know that this blockIter has in-place values.
-	return i.gatherBackward(&i.blockIter.ikey, base.MakeInPlaceValue(i.blockIter.val))
-}
-
-// SeekGE implements (keyspan.FragmentIterator).SeekGE.
-func (i *fragmentBlockIter) SeekGE(k []byte) (*keyspan.Span, error) {
-	if s, err := i.SeekLT(k); err != nil {
-		return nil, err
-	} else if s != nil && i.blockIter.cmp(k, s.End) < 0 {
-		return s, nil
-	}
-	// TODO(jackson): If the above i.SeekLT(k) discovers a span but the span
-	// doesn't meet the k < s.End comparison, then there's no need for the
-	// SeekLT to gatherBackward.
-	return i.Next()
-}
-
-// SeekLT implements (keyspan.FragmentIterator).SeekLT.
-func (i *fragmentBlockIter) SeekLT(k []byte) (*keyspan.Span, error) {
-	i.dir = -1
-	return i.gatherBackward(i.blockIter.SeekLT(k, base.SeekLTFlagsNone))
-}
-
-// String implements fmt.Stringer.
-func (i *fragmentBlockIter) String() string {
-	return "fragment-block-iter"
-}
-
-// SetCloseHook implements sstable.FragmentIterator.
-func (i *fragmentBlockIter) SetCloseHook(fn func(i keyspan.FragmentIterator) error) {
-	i.closeHook = fn
-}
-
-// WrapChildren implements FragmentIterator.
-func (i *fragmentBlockIter) WrapChildren(wrap keyspan.WrapFn) {}
